Share date range selector with usage page
